mulit-files-rag-agent-compare/requirements.txt
# LangChain
langchain_community==0.4.1
langchain_core==1.3.2
langchain_text_splitters==1.1.0
faiss-cpu==1.12.0

# LlamaIndex
llama-index-core==0.14.21
llama-index-llms-dashscope==0.6.1
llama-index-embeddings-dashscope==0.5.0
llama-index-readers-file==0.6.0
PyMuPDF==1.24.14

# DashScope
dashscope==1.22.2

# Qwen Agent
qwen_agent==0.0.34
